Edited some views and made everything work as expected
Now the application encrypts the password of the user in order to not store plain passwords text for a safer application. I have also implemented the pagination of the incidents list. The only thing left is to be able to locate the incidents in a map to visualize them way better.